Move tellstick base entity to separate module
