+++ homeassistant/components/tellstick/const.py
@@ -6,3 +6,5 @@ ATTR_DISCOVER_DEVICES = "devices"
 DATA_TELLSTICK = "tellstick_device"
 
 DEFAULT_SIGNAL_REPETITIONS = 1
+
+SIGNAL_TELLCORE_CALLBACK = "tellstick_callback"
